Full impl

+++ key.go
@@ -2,13 +2,13 @@ package main
 
 var keyChar = []byte("0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
 
-func genKey(n int) string {
+func genKey(n int64) string {
 	if n == 0 {
 		return string(keyChar[0])
 	}
-	l := len(keyChar)
+	l := int64(len(keyChar))
 	s := make([]byte, 20) // FIXME: will overflow. eventually.
-	i := len(s)
+	i := int64(len(s))
 	for n > 0 && i >= 0 {
 		i--
 		j := n % l
